requirements.txt
python-dotenv==0.13.0
requests==2.23.0

